+++ tests/test_server.py
@@ -31,6 +31,7 @@ def test_tools_list_has_read_tools_and_setup():
         "get_offline_access_points",
         "generate_health_report",
         "generate_inventory_report",
+        "generate_firmware_report",
         "find_client",
         "trace_client",
         "get_nac_clients",